+++ Hackerton/Assets/ChangeShapeMiniGame/Obstacle/Script/Obstacle.cs
@@ -24,7 +24,7 @@ namespace ChangeShapeMiniGame.Obstacle.Script
             if (other.CompareTag("Player") && Player.ShapeType != idNumber)
             {
                 Player.IsAlive = false;
-                GameOver.GoodGame();
+                UiSystem.GoodGame();
             }
 
             if (other.CompareTag("Wall"))
